+++ examples/actions/daytona_deactivate_snapshot/action.tf
@@ -1,0 +1,5 @@
+action "daytona_deactivate_snapshot" "example" {
+  config {
+    id = "snapshot-id"
+  }
+}